✨ 칭찬하기 화면 UI 구현
