+++ TwoToo/AppDelegate.swift
@@ -15,7 +15,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
 
-        KakaoSDK.initSDK(appKey: kakaoApiKey)
+//        KakaoSDK.initSDK(appKey: kakaoApiKey)
         Font.registerTTFont()
 
         // TODO: - info 파일 추가 시 주석 제거
